Broadcasting last message id and comparing message ids
The pull protocol requires to check if new messages are available or not. So we need to compare there last message ids.
Hence the broadcastLastMessageId() first fetches the last message id that is being received and then the message of type update is being broadcasted to connected devices. This update message is then processed in the received data subscription.

The compareMessageId compares the last received message id from the connected device with the last message id sent to the device. If they are not same, then messages in the cache are broadcasted.