Code formatting and UI fixes
The text editing controller had state issues due to placed in the build function.
The refreshMessages() function is formatted to make it empty before adding messages to it again.
scroll controller is used to make the controller place at the last position where the last message is lies.
Code is formatted according to Flutter updates and Listview of messages modified.