Code formatting, check for updates and UI fixed
The ListView is modified in the UI to fix render overflow.
checkDevices() function is separated from the init function to modularize the code. Also the devices auto connected. It is fixed by commenting the if statement at line 202.

init() calls the checkDevices() and calls broadcastLastMessageId(). It checks whether the message is type of update or not. If yes, then the compareMessageId function is called.